修改发送email

+++ company-service-provider/src/main/java/com/moseeker/company/service/impl/TalentpoolEmailService.java
@@ -1657,7 +1657,7 @@ public class TalentpoolEmailService {
                 for(Map<String,Object> app:applications){
                     int itemId=(int)app.get("company_id");
                     String title=(String)app.get("title");
-                    int status=Integer.parseInt(String.valueOf(app.get("status")));
+                    int status=(int)Double.parseDouble(String.valueOf(app.get("status")));
                     if(companyId==itemId&&status==0){
                         positionName=positionName+title+",";
                     }
@@ -1668,7 +1668,7 @@ public class TalentpoolEmailService {
                 for(Map<String,Object> app:applications){
                     int publisher=(int)app.get("publisher");
                     String title=(String)app.get("title");
-                    int status=Integer.parseInt(String.valueOf(app.get("status")));
+                    int status=(int)Double.parseDouble(String.valueOf(app.get("status")));
                     if(hrId==publisher&&status==0){
                         positionName=positionName+title+",";
                     }
